Added ability to add extra attributes (and related "guessers") to categories
Thus the line 'CategoryEvenInteger().AddGuesser('half: PSElement(magnitude=inst.magnitude / 2)')' adds
the ability to calculate 'half' (well, it actually *always* computes the half after that line, which is wrong).

The test works, but if flaky: not sure what the race condition is yet.

This will need refactor after some thought. The goal is to have something that gets added on demand, but can be reflected well in the LTM as well and will be more in control of the category itself.