Added files such as run_pyseqsee.py, to ready for end-to-end.
Until now, I have been building bottom up (getting into place objects, categories
relations, arena). This puts together some parts from the other end, such as
an intial version of GUI, ability to run on a ton of examples in batch, etc.

And  in the middle of those two, it adds starter code for focusing and the stream.

Next, work on codelets can commence.

What makes me unhappy: we currently seem to need to thread the "controller" argument in many places. This smells.