Add event "Setup finished".
This will allow up to add a lot of events before this point and know where the real action starts.
This is needed since we need to add categories to the history (they are, after all, the cause of a lot of action),
and a lot of them will be created at the start when we read from the LTM, inundating the history.

Added highlighting of what changed since previous refresh, so now it is easier to see what just happened.